+++ templates/mca/tests/MCA.IntegrationTests/Auth/ExternalProviderTests.cs
@@ -3,9 +3,8 @@ using System.Net;
 using System.Net.Http.Json;
 using System.Text.Json;
 using FluentAssertions;
-using Microsoft.AspNetCore.Hosting;
+using MCA.IntegrationTests;
 using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.Extensions.Configuration;
 using Xunit;
 
 namespace MCA.IntegrationTests.Auth;